requirements.txt
altair==5.4.0
attrs==24.2.0
blinker==1.8.2
branca==0.7.2
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
ckanapi==4.8
click==8.1.7
docopt==0.6.2
folium==0.17.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
narwhals==1.4.2
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.3
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
streamlit==1.37.1
streamlit_folium==0.22.0
tenacity==8.5.0
text-unidecode==1.3
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
xyzservices==2024.6.0
